implemented resources unpacking for android

+++ builds/android/build.sh
@@ -6,21 +6,25 @@ _JAVA_OPTIONS='-Dawt.useSystemAAFontSettings=lcd -Xms256m -Xmx512m -XX:MaxHeapSi
 ANDROID_NDK='/opt/android-ndk/'
 NDK_ROOT='/opt/android-ndk'
 
+BUILD_RESOURCES=1
+RES_DIR='res/raw'
+
 # FIXME
 cd ~/coding/github/ponic/builds/android
-
-#RES_DIR=res/raw
-RES_DIR=assets
-rm -fr $RES_DIR
 
 cp -r ../../src/* jni/
 rm -fr jni/platforms/linux/
 
-mkdir -p $RES_DIR
-mv jni/engine/shaders $RES_DIR/shaders
-cd $RES_DIR
-../../../tools/AtlasGen/AtlasGen.py -r ../../../resources -m RGBA -t tga
-cd ..
+if [[ $BUILD_RESOURCES == 1 ]]; then
+    rm -fr $RES_DIR
+    mkdir -p $RES_DIR
+    mv jni/engine/shaders $RES_DIR/shaders
+    cd $RES_DIR
+    ../../../../tools/AtlasGen/AtlasGen.py -r ../../../../resources \
+        -m RGBA -t tga
+    7z a res.zip atlases/ shaders/
+    cd ../..
+fi
 
 android update project --name ponic --path . --target "android-15"
 ndk-build V=1 -j4
